revert renaming of file
